Move kbuild-generic stuff into new package
This allows me to reuse all of this stuff when integrating u-boot.
Unfortunately, this also has been the final straw for removing the
initramfs CPIO archive from the build. It's difficult/not possible to
resolve paths from labels in macros, which makes it impossible to set
CONFIG_INITRAMFS_SOURCE correctly for the build.