Replace rules_nixpkgs with bazel-arm-gnu-toolchain
This makes a little bit of a mess of the kbuild rule, because it turns
out that it's not trivial to get a tool path from a CcToolchainInfo
provider--sometimes tool paths are not set, which requires the use of
cc_common.get_tool_for_action. I'll have to clean up the rule in a
future commit.